+++ open-codegen/opengen/templates/optimizer.rs
@@ -64,7 +64,7 @@ pub const {{meta.optimizer_name|upper}}_N2: usize = {{problem.dim_constraints_pe
 
 // ---Parameters of the constraints----------------------------------------------------------------------
 
-{% if 'Ball2' == problem.constraints.__class__.__name__ or 'BallInf' == problem.constraints.__class__.__name__ -%}
+{% if 'Ball1' == problem.constraints.__class__.__name__ or 'Ball2' == problem.constraints.__class__.__name__ or 'BallInf' == problem.constraints.__class__.__name__ -%}
 /// Constraints: Centre of Ball
 const CONSTRAINTS_BALL_XC: Option<&[f64]> = {% if problem.constraints.center is not none %}Some(&[{{problem.constraints.center | join(', ')}}]){% else %}None{% endif %};
 
@@ -110,7 +110,7 @@ const SET_C_XMAX :Option<&[f64]> = {% if problem.alm_set_c.xmax is not none %}So
 
 {% if problem.alm_set_y is not none -%}
 // ---Parameters of ALM-type constraints (Set Y)---------------------------------------------------------
-{% if 'Ball2' == problem.alm_set_y.__class__.__name__ or 'BallInf' == problem.alm_set_y.__class__.__name__ -%}
+{% if 'Ball1' == problem.alm_set_y.__class__.__name__ or 'Ball2' == problem.alm_set_y.__class__.__name__ or 'BallInf' == problem.alm_set_y.__class__.__name__ -%}
 /// Constraints: Centre of Euclidean Ball
 const SET_Y_BALL_XC: Option<&[f64]> = {% if problem.alm_set_y.center is not none %}Some(&[{{problem.alm_set_y.center | join(', ')}}]){% else %}None{% endif %};
 
@@ -136,6 +136,13 @@ fn make_constraints() -> impl Constraint {
     {% elif 'BallInf' == problem.constraints.__class__.__name__ -%}
     // - Infinity ball:
     BallInf::new(CONSTRAINTS_BALL_XC, CONSTRAINTS_BALL_RADIUS)
+    {% elif 'Ball1' == problem.constraints.__class__.__name__ -%}
+    // - Ball1:
+    Ball1::new(CONSTRAINTS_BALL_XC, CONSTRAINTS_BALL_RADIUS)
+    {% elif 'Simplex' == problem.constraints.__class__.__name__ -%}
+    // - Simplex:
+    let alpha_simplex : f64 = {{problem.constraints.alpha}};
+    Simplex::new(alpha_simplex)
     {% elif 'Rectangle' == problem.constraints.__class__.__name__ -%}
     // - Rectangle:
     Rectangle::new(CONSTRAINTS_XMIN, CONSTRAINTS_XMAX)
@@ -165,11 +172,20 @@ fn make_constraints() -> impl Constraint {
         let radius_{{loop.index}} = {{set_i.radius}};
         let center_{{loop.index}}: Option<&[f64]> = {% if set_i.center is not none %}Some(&[{{set_i.center | join(', ')}}]){% else %}None{% endif %};
         let set_{{loop.index}} = Ball2::new(center_{{loop.index}}, radius_{{loop.index}});
-        let bounds = bounds.add_constraint(idx_{{loop.index}}, set_{{loop.index}});
+        let bounds = bounds.add_constraint(idx_{{loop.index}}, set_{{loop.index}});        
         {% elif 'BallInf' == set_i.__class__.__name__ -%}
         let radius_{{loop.index}} = {{set_i.radius}};
         let center_{{loop.index}}: Option<&[f64]> = {% if set_i.center is not none %}Some(&[{{set_i.center | join(', ')}}]){% else %}None{% endif %};
         let set_{{loop.index}} = BallInf::new(center_{{loop.index}}, radius_{{loop.index}});
+        let bounds = bounds.add_constraint(idx_{{loop.index}}, set_{{loop.index}});
+        {% elif 'Ball1' == set_i.__class__.__name__ -%}
+        let radius_{{loop.index}} = {{set_i.radius}};
+        let center_{{loop.index}}: Option<&[f64]> = {% if set_i.center is not none %}Some(&[{{set_i.center | join(', ')}}]){% else %}None{% endif %};
+        let set_{{loop.index}} = Ball1::new(center_{{loop.index}}, radius_{{loop.index}});
+        let bounds = bounds.add_constraint(idx_{{loop.index}}, set_{{loop.index}});
+        {% elif 'Simplex' == set_i.__class__.__name__ -%}
+        let alpha_{{loop.index}} = {{set_i.alpha}};        
+        let set_{{loop.index}} = Simplex::new(alpha_{{loop.index}});
         let bounds = bounds.add_constraint(idx_{{loop.index}}, set_{{loop.index}});
         {% elif 'Rectangle' == set_i.__class__.__name__ -%}
         let xmin_{{loop.index}} :Option<&[f64]> = {% if set_i.xmin is not none %}Some(&[
@@ -209,6 +225,11 @@ fn make_set_c() -> impl Constraint {
     Ball2::new(SET_C_BALL_XC, SET_C_BALL_RADIUS)
     {% elif 'BallInf' == problem.alm_set_c.__class__.__name__ -%}
     BallInf::new(SET_C_BALL_XC, SET_C_BALL_RADIUS)
+    {% elif 'Ball1' == problem.alm_set_c.__class__.__name__ -%}
+    Ball1::new(SET_C_BALL_XC, SET_C_BALL_RADIUS)
+    {% elif 'Simplex' == problem.alm_set_c.__class__.__name__ -%}
+    let set_c_simplex_alpha : f64 = {{problem.alm_set_y.alpha}};
+    Simplex::new(set_c_simplex_alpha)
     {% elif 'Rectangle' == problem.alm_set_c.__class__.__name__ -%}
     Rectangle::new(SET_C_XMIN, SET_C_XMAX)
     {% elif 'NoConstraints' == problem.alm_set_c.__class__.__name__ -%}
@@ -230,6 +251,15 @@ fn make_set_c() -> impl Constraint {
             let radius_{{loop.index}} = {{set_i.radius}};
             let center_{{loop.index}}: Option<&[f64]> = {% if set_i.center is not none %}Some(&[{{set_i.center | join(', ')}}]){% else %}None{% endif %};
             let set_{{loop.index}} = BallInf::new(center_{{loop.index}}, radius_{{loop.index}});
+            let set_c = set_c.add_constraint(idx_{{loop.index}}, set_{{loop.index}});
+            {% elif 'Ball1' == set_i.__class__.__name__ -%}
+            let radius_{{loop.index}} = {{set_i.radius}};
+            let center_{{loop.index}}: Option<&[f64]> = {% if set_i.center is not none %}Some(&[{{set_i.center | join(', ')}}]){% else %}None{% endif %};
+            let set_{{loop.index}} = Ball1::new(center_{{loop.index}}, radius_{{loop.index}});
+            let set_c = set_c.add_constraint(idx_{{loop.index}}, set_{{loop.index}});
+            {% elif 'Simplex' == set_i.__class__.__name__ -%}
+            let alpha_smplx_{{loop.index}} = {{set_i.alpha}};
+            let set_{{loop.index}} = Simplex::new(alpha_smplx_{{loop.index}});
             let set_c = set_c.add_constraint(idx_{{loop.index}}, set_{{loop.index}});
             {% elif 'Rectangle' == set_i.__class__.__name__ -%}
             let xmin_{{loop.index}} :Option<&[f64]> = {% if set_i.xmin is not none %}Some(&[