+++ CHANGELOG.md
@@ -12,7 +12,15 @@ Note: This is the main Changelog file for the Rust solver. The Changelog file fo
       Unreleased
      --------------------- -->
 
+<!-- ---------------------
+      v0.7.3
+     --------------------- -->     
+## [v0.7.3] - 2021-11-1
 
+### Added 
+
+* Implementation of Simplex and Ball1 constraints in Rust
+* Fix issue with simultaneous use of features `jem` and `rp`
 
 
 <!-- ---------------------
@@ -198,7 +206,8 @@ This is a breaking API change.
      --------------------- -->
 
 <!-- Releases -->
-[Unreleased]:: https://github.com/alphaville/optimization-engine/compare/v0.7.2...master 
+[Unreleased]:: https://github.com/alphaville/optimization-engine/compare/v0.7.3...master 
+[v0.7.3]: https://github.com/alphaville/optimization-engine/compare/v0.7.2...v0.7.3
 [v0.7.2]: https://github.com/alphaville/optimization-engine/compare/v0.7.1...v0.7.2
 [v0.7.1]: https://github.com/alphaville/optimization-engine/compare/v0.7.0...v0.7.1
 [v0.7.0]: https://github.com/alphaville/optimization-engine/compare/v0.6.2...v0.7.0